Performance updates and code cleanup (#849)

+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/DropTableNode.java
@@ -65,7 +65,7 @@ public class DropTableNode extends DDLStatementNode
 		throws StandardException
 	{
 		initAndCheck(dropObjectName);
-		this.dropBehavior = ((Integer) dropBehavior).intValue();
+		this.dropBehavior = (Integer) dropBehavior;
 	}
 
 	/**
